### 2025-04-02 #7993 @editorialbot
[PRE REVIEW]: `pyScienceMode`: an Open-Source Python Package to control electro-stimulator through the Hasomed’s science mode protocol
**Submitting author:** @Kev1CO (<a href="http://orcid.org/0009-0009-0248-3548">Kevin Co</a>)
**Repository:** https://github.com/s2mLab/pyScienceMode
**Branch with paper.md** (empty if default branch): main
**Version:** 1.1.4
