### 2025-04-02 1caec15
doi added to references and joss submitted badge
