### 2025-05-17 #7993 @sappelhoff
Thanks for the tag @Kevin-Mattheus-Moerman, but before I commit to anything, I'd first like to clarify a few points.

Given that this submission is software for interacting with specific hardware, I think it would be good to find at least one peer reviewer with access to this hardware, who could actually test the software. @Kev1CO do you think we could find such a peer reviewer, especially considering the [JOSS COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)? Who would that be (a list of people would of course be best)?

And another question: What are your future plans for the software (1 year, 5 years, ...)?
