### 2025-05-21 #7993 @Kev1CO
Dear @sappelhoff,
I found 2 peers, Mr. Aksöz Efe Anil (efeanil.aksoez@bfh.ch) and Mr. Lecompte Olivier (olivier.lecompte@polymtl.ca) that kindly accepted to review the package, and they possess the stimulators.
This submission being my first JOSS contribution, what are the next steps to start the reviewing process ?
