### 2025-05-21 #7993 @sappelhoff
@Kev1CO thanks a lot for the initial inquiries, I'll be editing this submission.

Do your two peers have GitHub accounts? If so, please share the handles with me. If not, we would need them to create them, as JOSS submissions are reviewed openly and transparently on GitHub.
