### 2025-05-22 #7993 @sappelhoff
Thanks @efeanilaksoz!

> Next week, I plan to officially begin the review process. Please let me know if there’s anything else I should take into account beforehand.

The usual order of events is that in this current thread (the "PRE-review"), we assign an editor, find suitable peer reviewers (usually 2 to 3), and discuss initial issues, should they arise.

We now have an editor (myself) and one peer-reviewer assigned (you), and are waiting for at least one more (hopefully Mr. Lecompte Olivier).

Once we have the final peer-reviewer, I will start the actual review in a new thread and will make sure to inform you. Please do not start the review before I have assigned the remaining peer reviewer and have started the actual review thread 🙂 thanks!
