### 2025-05-23 #8259 @sappelhoff
@Kev1CO did you recently make changes to the paper? It was building fine before, see: https://github.com/openjournals/joss-reviews/issues/7993#issuecomment-2772309885

However in the current thread, there seems to be an error, see: https://github.com/openjournals/joss-reviews/issues/8259#issuecomment-2903706158

Please try to fix the error using the JOSS resources (https://joss.readthedocs.io/en/latest/index.html), and then try to use `@editorialbot` to rebuild the paper. Thanks!
