+++ docs/paper.bib
@@ -6,7 +6,7 @@
   number={2},
   pages={e0149024},
   year={2016},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
   doi={10.1371/journal.pone.0149024},
 }
 
@@ -18,7 +18,7 @@
   number={3},
   pages={738--749},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/TBME.2019.2920346},
 }
 
@@ -30,7 +30,7 @@
   volume={8},
   pages={711388},
   year={2021},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
   doi={10.3389/frobt.2021.711388},
 }
 
@@ -42,7 +42,7 @@
   number={4},
   pages={634--641},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/mus.25777},
 }
 
@@ -50,7 +50,7 @@
   title={A LabVIEW Interface for RehaStim 2},
   author={RaviChandran, Narrendar and Aw, Kean and McDaid, Andrew},
   year={2022},
-  publisher={TechRxiv}
+  publisher={TechRxiv},
   doi={10.36227/techrxiv.21302865.v1},
 }
 
@@ -61,6 +61,6 @@
   volume={8},
   number={83},
   pages={5091},
-  year={2023}
+  year={2023},
   doi={10.21105/joss.05091},
 }
