### 2025-06-17 #8259 @efeanilaksoz
> [@efeanilaksoz](https://github.com/efeanilaksoz) do you need any help with getting started with your review? Could you please update us on your plans on completing the review?

Dear all, I have everything I need as well. I plan to submit my review until next week Friday (27.06)
