### 2025-06-17 #8259 @sappelhoff
Thanks @efeanilaksoz please make sure to generate your review checklist!

You can read more about that in this comment (see above): https://github.com/openjournals/joss-reviews/issues/8259#issue-3085643166
