### 2025-06-19 #8259 @sappelhoff
Hi @olilecompte and @Kev1CO please make sure all review points end up in this thread, or in issues/pull requests on the submitted repo (https://github.com/s2mLab/pyScienceMode). This way we can have a transparent and traceable review, which is what we aim for at JOSS 🙂

If it is a list of miscellaneous comments, I think the best place would be as part of a comment in the present thread. Specific issues are best raised in the target repo, and linked back here by placing the link to this thread (https://github.com/openjournals/joss-reviews/issues/8259) into the text on the target repo.
