+++ examples/perf_test.py
@@ -1,5 +1,5 @@
 from pysciencemode import Channel, Point, Device, Modes
-from pysciencemode import RehastimP24 as St
+from pysciencemode import P24 as St
 from pysciencemode import Rehastim2 as St2
 import random
 from time import sleep
@@ -11,7 +11,7 @@ from biosiglive import ViconClient, DeviceType
 import numpy as np
 
 """
-This file is used to test the performance of both devices (RehastimP24 and Rehastim2).
+This file is used to test the performance of both devices (P24 and Rehastim2).
 The tests are done with Vicon Nexus and the biosiglive library.
 Some tests can be used for both devices, some others are specific to one device. You can
 choose the device you want by putting the right device enum in the function.
@@ -50,7 +50,7 @@ def init_trigger():
 
 def custom_shape_pulse():
     """
-    Test of the new feature of the RehastimP24 : custom shape pulse
+    Test of the new feature of the P24 : custom shape pulse
     """
     stimulatorp24 = St(port="COM4", show_log=True)
     channel_1 = Channel(
@@ -58,7 +58,7 @@ def custom_shape_pulse():
         name="Biceps",
         amplitude=30,
         frequency=20,
-        device_type=Device.Rehastimp24,
+        device_type=Device.P24,
     )
 
     list_channels.append(channel_1)
@@ -99,7 +99,7 @@ def single_doublet_triplet(device: Device):
     """
     Mode test (single, doublet, triplet) for both devices. It uses 3 channels at the same time.
     """
-    if device == Device.Rehastimp24:
+    if device == Device.P24:
         stimulatorp24 = St(port="COM4", show_log=True)
         channel_1 = Channel(
             mode=Modes.SINGLE,
@@ -108,7 +108,7 @@ def single_doublet_triplet(device: Device):
             amplitude=30,
             pulse_width=350,
             frequency=20,
-            device_type=Device.Rehastimp24,
+            device_type=Device.P24,
         )
         channel_2 = Channel(
             mode=Modes.DOUBLET,
@@ -117,7 +117,7 @@ def single_doublet_triplet(device: Device):
             amplitude=30,
             frequency=20,
             pulse_width=350,
-            device_type=Device.Rehastimp24,
+            device_type=Device.P24,
         )
         channel_3 = Channel(
             mode=Modes.TRIPLET,
@@ -126,7 +126,7 @@ def single_doublet_triplet(device: Device):
             amplitude=30,
             frequency=20,
             pulse_width=350,
-            device_type=Device.Rehastimp24,
+            device_type=Device.P24,
         )
 
         list_channels.append(channel_1)
@@ -212,7 +212,7 @@ def frequency_test(device: Device):
     Test of the frequency for both devices,  we compare with the real frequency measured with Vicon and pyomeca.
     """
     # Test with 10Hz
-    if device == Device.Rehastimp24:
+    if device == Device.P24:
         stimulatorp24 = St(port="COM4", show_log=True)
         channel_1 = Channel(
             mode=Modes.SINGLE,
@@ -221,7 +221,7 @@ def frequency_test(device: Device):
             amplitude=30,
             pulse_width=350,
             frequency=10,
-            device_type=Device.Rehastimp24,
+            device_type=Device.P24,
         )
         list_channels.append(channel_1)
 
@@ -277,7 +277,7 @@ def frequency_test(device: Device):
 
 def force_rehastim2():
     """
-    Same pattern as the force_rehastimp24 to compare the output force of the ergometer handle of the two devices
+    Same pattern as the force_p24 to compare the output force of the ergometer handle of the two devices
     """
     stimulator2 = St2(port="COM3", show_log=True)
[…]
