+++ README.md
@@ -1,10 +1,10 @@
 # pyScienceMode
-`pyScienceMode` is a Python interface to control the Rehastim2 and the RehastimP24. 
+`pyScienceMode` is a Python interface to control the Rehastim2 and the P24. 
 Please have a look to the documentation for more information about [pyScienceMode](https://pysciencemode.readthedocs.io/en/latest/index.html)
 
 ## How to install
 These are the different ways to install pyScienceMode for the Rehastim2 control.
-To control the RehastimP24, please follow the given steps in `Installing from source` as it requires an additional wheel compilation.
+To control the P24, please follow the given steps in `Installing from source` as it requires an additional wheel compilation.
 
 ### Installing from PyPI
 ```bash
@@ -20,11 +20,11 @@ conda install -c conda-forge pysciencemode
 Please refer to the [documentation](https://pysciencemode.readthedocs.io/en/latest/install.html) to install pyScienceMode. 
 
 ## How to use
-A set of example is provided in the `examples` folder to show how to control the Rehastim2 and the RehastimP24 :
+A set of example is provided in the `examples` folder to show how to control the Rehastim2 and the P24 :
 Please take a look at the [documentation example page](https://pysciencemode.readthedocs.io/en/latest/examples.html) for description of each example.
 
-## Main differences between the Rehastim2 and the RehastimP24
-They are some differences between the Rehastim2 and the RehastimP24.
+## Main differences between the Rehastim2 and the P24
+They are some differences between the Rehastim2 and the P24.
 Please take a look at the [documentation main differences page](https://pysciencemode.readthedocs.io/en/latest/main_differences.html) for more information.
 
 ## How to contribute
