+++ docs/main_differences.rst
@@ -1,5 +1,5 @@
 ==================================================
-Main differences between Rehastim2 and RehastimP24
+Main differences between Rehastim2 and P24
 ==================================================
 
 .. image:: /_static/tab_main_differences.png
