+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 [project]
 name = "pysciencemode"
 version = "1.1.4"
-description = "Python interface to control the Rehastim2 and RehastimP24 devices."
+description = "Python interface to control the Rehastim2 and P24 devices."
 readme = "README.md"
 license = { text = "MIT" }
 authors = [
