### 2025-06-19 1b0f3d8
Modification to reviewer comment on state of field
