### 2025-06-19 #8259 @Kev1CO
## Answer to comments from reviewer olilecompte

Dear @olilecompte, thank you for your review.
Please tell me if the provided answers and modifications corresponds to those requested.

### Functionality:
#### Installation:

Reviewer:

> The installation is very well documented and easy to follow. It is actually much easier to follow this documentation than the one provided by Hasomed.

Authors:

> This is appreciated.

#### Functionality:
Reviewer:

> I was able to verify most of the claims regarding acks, channel configuration, and overall functionality with the P24 stimulator. I also ran the example for the stimulator I own and everything worked perfectly. However, I cannot verify the switching between stimulators nor the real-time communication with the Motomed.

Authors:

> How should we proceed to validate that @sappelhoff? I did find reviewers that possess the P24 stimulator but did not ask if they owned the Rehastim2 device nor the MOTOmed bike mentioned in the paper. Shall we wait @efeanilaksoz’s verification, if he possesses the 2 devices. Or make a video to demonstrate it works or a visio conference meeting…?

### Documentation:
#### Statement of need:
Reviewer:

> The statement of need is not explicit in the documentation.

Authors:

> We added this paragraph to the GitHub readme in commit #Statement of need:

> _Functional electrical stimulation (FES) research would benefit of an open, flexible control method for customizable stimulation patterns. pyScienceMode provides a unified Python API for both Rehastim2 and P24 devices, enabling real-time adjustment of frequency, intensity, pulse width, train duration and sensor-triggered control. It supports rapid prototyping of personalized, real-time FES protocols, making novel rehabilitation strategies reproducible, adaptable and easily extensible as new hardware emerges._

#### Automated tests:
Reviewer:

> There are no automated tests.

Authors:

> Tests are available in the tests file. Unfortunately, because a device is required to run them, the tests were not performed online (GitHub action). But they are still runnable offline when adding new commits and pull requests.

### Software paper:
#### State of field:
Reviewer :

> The authors do not address how their software compares to the Python wrapper provided by Hasomed (https://github.com/ScienceMode/ScienceMode4_python_wrapper). This is arguably the most relevant point of comparison, yet it is not mentioned. It provides most of the core functionalities highlighted in the current paper and is capable of addressing most, if not all, of the problems outlined in the statement of need. It is also possible to integrate it with other sensors via multiprocessing, which is therefore not a specific "key advantage" of the pyScienceMode software.

Authors :

> Indeed, the python wrapper processes the same functionalities than pyScienceMode so the following modifications to the paper were made to justify the relevance of pyScienceMode in commit #Modification to reviewer comment on state of field:

> _While a python wrapper is available for the P24 stimulator (https://github.com/ScienceMode/ScienceMode4_python_wrapper), already enabling customized control, none are available for the Rehastim2 stimulator.  pyScienceMode regroups the functionalities of both the Rehastim2 and P24 stimulators, providing a unified interface, a user-friendly coding environment and a simple installation procedure._

> And the « key advantage » term was replaced by:

> _It can also integrate other sensors and devices…_

#### References:
Reviewer:

> The first reference is not cited using the proper citation syntax.

Authors:

> Typo error was corrected in #commit reference typo

### Varia
#### General comment:
Reviewer:

> I believe the proper name for the newest stimulator is the P24 and not the RehastimP24 (from the brochure). I think the name should be modified to reflect this in the documentation and the codes.

Authors:

> Modification made from RehastimP24 to P24 in commit #Modifying name RehastimP24 device to P24
