### 2025-06-20 #8259 @olilecompte
@Kev1CO  Thank you for the revisions. I feel that most of the comments have been adequately addressed.

My only remaining concern relates to the scope of the paper, particularly in relation to the statement of need and prior work. If the core contribution of pyScienceMode lies in its unified support for both the Rehastim2 and P24 stimulators, this should be stated more clearly. As it stands—even with the revisions—the contribution remains somewhat vague, and the comparison to existing tools still lacks transparency.
