### 2025-06-20 #8259 @Kev1CO
@olilecompte, I understand your concerns. The main contribution relies on the possibility to control the Rehastim2 from a python framework, which has never been possible in open-source (mentioned in the new version of the article) and we did not say it was better than other methods (no performance comparison). As for the P24 stimulator, it is just an overlay using the python_p24_wrapper (as rectified in the new version of the article), so comparison between methods is not relevant as both are the same. It is just to control it, alike the Rehastim2 stimulator we have developed.

If I replace:
> While a python wrapper is available for the P24 stimulator (https://github.com/ScienceMode/ScienceMode4_python_wrapper), already enabling customized control, none are available for the Rehastim2 stimulator. pyScienceMode regroups the functionalities of both the Rehastim2 and P24 stimulators, providing a unified interface, a user-friendly coding environment and a simple installation procedure.

for:
> pyScienceMode is the first open-source package enabling the customization of stimulation patterns for the Rehastim2 stimulator in Python. The P24 stimulator control was also included to the package (based on https://github.com/ScienceMode/ScienceMode4_python_wrapper) to provide a unified interface and a user-friendly coding environment, accessible from a simple installation procedure.

does it answer your concerns ?
