### 2025-06-25 acc09a4
Enhance port connection
