### 2025-06-27 71c9978
How to use pysciencemode
