### 2025-06-27 6e24f2f
Additional installation step for p24
