### 2025-06-27 f57b49a
paper modification
