### 2025-06-27 #8259 @Kev1CO
@olilecompte I updated the paper with the proposed modifications

> pyScienceMode is the first open-source package enabling the customization of stimulation patterns for the Rehastim2 stimulator in Python. The P24 stimulator control was also included to the package (based on https://github.com/ScienceMode/ScienceMode4_python_wrapper) to provide a unified interface and a user-friendly coding environment, accessible from a simple installation procedure.
