### 2025-06-27 #8259 @Kev1CO
@efeanilaksoz

### Installation:
Reviewer:
> The installation process worked smoothly on Windows platform. However, I encountered an issue on macOS (this could be my personal mistake) but if it's not:
>
> ERROR: sciencemode_cffi-1.0.0-cp310-cp310-win_amd64.whl is not a supported wheel on this platform.
>
> The provided wheel file (sciencemode_cffi-1.0.0-cp310-cp310-win_amd64.whl) is Windows-specific?
>
> To improve cross-platform compatibility, please consider:
>
> Updating the installation guide to explicitly state that the current .whl file is intended for Windows only, and that users on other platforms will need an alternative approach.
> Adding a note regarding the plan for providing a platform-independent wheel or clear instructions for building sciencemode_cffi from source on macOS and Linux.
> Including these details would significantly enhance the user experience for non-Windows environments.

Authors:
> Indeed, macOS is not supported by the wheel given in pysciencemode. This is because the original wrapper (https://github.com/ScienceMode/ScienceMode4_python_wrapper) does not enable macOS compatibility. Unfortunately, we do not have the capability to solve the compatibility issue, but if the Hasomed ScienceMode enables it, we will add the possibility to do so in our package. Regarding this, I opened an issue on the wrapper repository (https://github.com/ScienceMode/ScienceMode4_python_wrapper/issues/5).

> To improve the installation guide, the following information was added to the readme file commit #additional installation step for the p24:
> Additional installation step for P24 :
> For Python 3.10 only : navigate to the folder where the file sciencemode_cffi-1.0.0-cp310-cp310-win_amd64.whl is located and run the following command :
> pip install sciencemode_cffi-1.0.0-cp310-cp310-win_amd64.whl
> For other Python version : build your own wheel from https://github.com/ScienceMode/ScienceMode4_python_wrapper and pip install it in your environment.

### Functionality
Reviewer:
> All the P24 science functions worked without any problem. We don't have Rehastim2 so can't verify anything about Rehastim2. My review and these comments are explicitly for P24 science.
>
> The provided code does not explicitly check if the specified serial port (e.g., COM# or '/dev/tty.usbmodemXXX') is available before initializing the device connection. Nor does it list available ports in case of failure.
>
> I believe the code uses In init, it calls the super().init(port, ...), which delegates the port handling to RehastimGeneric (assumed to handle the low-level serial communication setup). If the port is invalid or unavailable, a connection error (e.g., OSError, SerialException) would likely be raised there, but it is not explicitly caught or handled within RehastimP24.
>
> It will be beneficial for the users if it checks the list of available ports before attempting a connection. It can provide a user-friendly error message if the port is invalid. It can also import or use utilities like serial.tools.list_ports to enumerate available COM ports (which could help guide the user).
>
> Suggested Improvement: Adding a utility function before initialization to validate the port.

Authors:
> Thank you for this feedback, we addressed this functionality in commit #Enhance port connection
> The new version gives the available ports to the user if the entered port is not correct. If the user entered an existing port but is not recognized as the stimulator, it will give the list of other potential ports before leading to error: a connection error (e.g., OSError, SerialException). I did not make an automatic connection to be as flexible as possible in the port chose to allow user to control two or more stimulators simultaneously with the same computer.

### Documentation
Reviwer:
> It will be very beneficial to include P24_Instructions_for_Use_v1.1_20221206[12482].pdf file or new versions. If the authors have an access and of course if it is agreed by Hasomed. In documentation, the link to Hasomed link can also include a link to the instructions for use if the website provides.
>
> Refering to the P24 instructions for use document, it should be stated that:
> The P24 Science/P24 Module is a device that can be controlled by a computer system via a specified interface to generate and output electrical pulses. The P24 Science/P24 Module is intended for research applications only and is not intended to be used for medical purposes on human beings according to Regulation (EU) 2017/745.

Authors:
> After contacting Hasomed to see if they would agree to let user guide documents available in pysciencemode repository, we added the links to the readme file and the regulation concern in commit #instruction for use
