### 2025-06-30 #8259 @sappelhoff
@Kev1CO please work on your points from https://github.com/openjournals/joss-reviews/issues/8259#issuecomment-3019616057 and let me know once you are done.

I will then continue with my list.
