### 2025-07-01 #8259 @Kev1CO
Thank you very much!
Sorry, I have limited access to internet this week.
I will address the remaining checklist on Monday.
