+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "pysciencemode"
-version = "1.1.4"
+version = "1.1.5"
 description = "Python interface to control the Rehastim2 and P24 devices."
 readme = "README.md"
 license = { text = "MIT" }
