### 2025-07-10 66b8454
abbreviation, link and correction
