### 2025-07-25 #8259 @Kev1CO
Hi @Kevin-Mattheus-Moerman !
I have noticed that on the Google Scholar search : "pysicencemode", my name does not appear as an author whereas in the joss journal I do. Any idea on how to fix that ?
Thank you very much for your help !
