### 2025-09-03 #8259 @arfon
@Kev1CO – I'm afraid there's nothing we can do to control Google Scholar's behavior. The metadata looks correct on the JOSS site, and that's all the control we have.

```

<meta name="citation_title" content="&lt;code&gt;pyScienceMode&lt;/code&gt;: an Open-Source Python Package to control electro-stimulator through the Hasomed’s science mode protocol">
<meta name="citation_author" content="Kevin Co">
<meta name="citation_author" content="Amedeo Ceglia">
<meta name="citation_author" content="Mickael Begon">

<meta name="citation_publication_date" content="2025/07/11">
<meta name="citation_journal_title" content="Journal of Open Source Software">
<meta name="citation_doi" content="10.21105/joss.08259">
<meta name="citation_issn" content="2475-9066">
<meta name="citation_volume" content="10">
<meta name="citation_issue" content="111">
<meta name="citation_firstpage" content="8259">
<meta name="citation_abstract_html_url" content="https://joss.theoj.org/papers/10.21105/joss.08259">
<meta name="citation_pdf_url" content="https://joss.theoj.org/papers/10.21105/joss.08259.pdf">

```
